Moved fields files metaboxes folder to src
