+++ src/Fields/AbstractInput.php
@@ -21,11 +21,11 @@
  * @subpackage WPAbstractClasses
  * @author     Kevin Roy <royk@myraytech.net>
  * @license    GPL-v2 <https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html>
- * @version    0.1.0
+ * @version    0.2.0
  * @since      0.1.0
  */
 
-namespace RayTech\WPAbstractClasses\MetaBoxes\Fields;
+namespace RayTech\WPAbstractClasses\Fields;
 
 /**
  * Abstract input class
