+++ src/Fields/Inputs/Number.php
@@ -21,18 +21,18 @@
  * @subpackage WPAbstractClasses
  * @author     Kevin Roy <royk@myraytech.net>
  * @license    GPL-v2 <https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html>
- * @version    0.1.0
+ * @version    0.2.0
  * @since      0.1.0
  */
 
-namespace RayTech\WPAbstractClasses\MetaBoxes\Fields\Inputs;
+namespace RayTech\WPAbstractClasses\Fields\Inputs;
 
-use RayTech\WPAbstractClasses\MetaBoxes\Fields\AbstractInput;
+use RayTech\WPAbstractClasses\Fields\AbstractInput;
 
 /**
- * DateTime input
+ * Number input
  */
-class DateTime extends AbstractInput {
+class Number extends AbstractInput {
 	/**
 	 * __construct
 	 *
@@ -44,9 +44,9 @@ class DateTime extends AbstractInput {
 	 * @return void
 	 */
 	public function __construct( $id, $name, $value, $attr ) {
-		$this->setType( 'date' );
-		$this->setName( $name );
+		$this->setType( 'number' );
 		$this->setInputID( $id );
+		$this->setName( $name );
 		$this->setValue( $value );
 		$this->setAttributes( $attr );
 	}
