+++ src/Fields/Inputs/Select.php
@@ -21,18 +21,18 @@
  * @subpackage WPAbstractClasses
  * @author     Kevin Roy <royk@myraytech.net>
  * @license    GPL-v2 <https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html>
- * @version    0.1.0
+ * @version    0.2.0
  * @since      0.1.0
  */
 
-namespace RayTech\WPAbstractClasses\MetaBoxes\Fields\Inputs;
+namespace RayTech\WPAbstractClasses\Fields\Inputs;
 
-use RayTech\WPAbstractClasses\MetaBoxes\Fields\AbstractInput;
+use RayTech\WPAbstractClasses\Fields\AbstractInput;
 
 /**
- * Media input
+ * Select input
  */
-class Media extends AbstractInput {
+class Select extends AbstractInput {
 	/**
 	 * __construct
 	 *
@@ -44,41 +44,39 @@ class Media extends AbstractInput {
 	 * @return void
 	 */
 	public function __construct( $id, $name, $value, $attr ) {
-		$this->setType( 'media' );
-		$this->setInputID( $id );
 		$this->setName( $name );
+		$this->setInputID( $id );
 		$this->setValue( $value );
 		$this->setAttributes( $attr );
 	}
 
 	/**
-	 * Rendering method
+	 * Rendering method.
 	 *
-	 * @access public
 	 * @return void
 	 */
 	public function render() {
-		{
-			$image            = ' button">Upload image';
-			$image_size       = 'full'; // It would be better to use thumbnail size here (150x150 or so).
-			$display          = 'none'; // Display state ot the "Remove image" button.
-			$image_attributes = wp_get_attachment_image_src( $this->getValue(), $image_size );
-
-		if ( $image_attributes ) {
-
-			// $image_attributes[0] - image URL
-			// $image_attributes[1] - image width
-			// $image_attributes[2] - image height
-
-			$image   = '"><img src="' . esc_url( $image_attributes[0] ) . '" style="max-width:100px;display:block;" />';
-			$display = 'inline-block';
+		echo '<select id="' . esc_attr( $this->getInputId() ) . '" name="' . esc_attr( $this->getName() ) . '"';
+		if ( ! empty( $this->getAttributes() ) ) {
+			foreach ( $this->getAttributes() as $attr => $attr_value ) {
+				if ( 'options' === $attr ) {
+					continue;
+				} elseif ( 'multiple' === $attr && true === $attr_value ) {
+					echo ' multiple';
+				} else {
+					echo ' ' . esc_html( $attr ) . '="' . esc_attr( $attr_value ) . '"';
+				}
+			}
 		}
+		echo '>';
 
-		echo '<div>
-                <a href="#" id="' . esc_attr( $this->getInputId() ) . 's-image" class="' . esc_attr( THEME_NAME ) . '_upload_image_button' . $image . '</a>
-                <input type="hidden" name="' . esc_attr( $this->getName() ) . '" id="' . esc_attr( $this->getInputId() ) . '" value="' . esc_attr( $this->getValue() ) . '" />
-                <a href="#" class="' . esc_attr( THEME_NAME ) . '_remove_image_button" style="display:' . esc_attr( $display ) . '">Remove image</a>
-            </div>';
+		foreach ( $this->getAttributes()['options'] as $option => $label ) {
+			echo '<option value="' . esc_attr( $option ) . '"';
+			if ( $this->getValue() === $option ) {
+				echo ' selected';
+			}
+			echo '>' . esc_html( $label ) . '</option>';
 		}
+		echo '</select>';
 	}
 }
